# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BOND-GYARADOS-API",
    "description": "En esta documentacion encontrara toda la informacion relacionada con la adquicion de informacion para el proyecto bond-gyarados-api y su ulizacion.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://<NameResource>.azurewebsites.net/api/<NameFunction>/<NameOperation>"
    }
  ],
...
  "paths": {
...
    "/api/organization/add": {
      "post": {
        "tags": [
          "organization"
        ],
        "summary": "Operation to add organization",
        "description": "post",
        "operationId": "postOrganizationAdd",
        "requestBody": {
          "description": "Operation  to add organization",
          "content": {
            "Add Primary Group": {
              "schema": {
                "$ref": "#/components/schemas/AddPrimaryGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseAddPrimaryGroup"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/organization/update": {
      "post": {
        "tags": [
          "organization"
        ],
        "summary": "Operation to update organization",
        "operationId": "postOrganizationUpdate",
        "requestBody": {
          "description": "Operationto update organization",
          "content": {
            "Update Primary Group": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrimaryGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUpdatePrimaryGroup"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/organization/MainPage": {
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "get info for organization",
        "description": "get info for organization",
        "operationId": "getorganization",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Value for organization",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "MainPage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseConsultOrganization"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/organization/delete": {
      "delete": {
        "tags": [
          "organization"
        ],
        "summary": "Operation to delete organization",
        "operationId": "postOrganizationDelete",
        "parameters": [
          {
            "name": "NamePrimary",
            "in": "query",
            "description": "Value for user",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "operator"
            }
          },
          {
            "name": "confirmation",
            "in": "query",
            "description": "Confirmation delete",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDeleteGroup"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/uplinks/add": {
      "post": {
        "tags": [
          "uplinks"
        ],
        "summary": "Operation about uplinks",
        "operationId": "postuplinks",
        "requestBody": {
          "description": "Operation about uplinks",
          "content": {
            "Add Uplinks": {
              "schema": {
                "$ref": "#/components/schemas/AddUplinks"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUplinks"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/uplinks/tableValues": {
      "get": {
        "tags": [
          "uplinks"
        ],
        "summary": "get info for uplinks",
        "description": "get info for uplinks",
        "operationId": "getuplinks",
        "parameters": [
          {
            "name": "Entity",
            "in": "query",
            "description": "Value for entity",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "3ra. Bomba"
            }
          },
          {
            "name": "State",
            "in": "query",
            "description": "Value for state",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "activate"
            }
          },
          {
            "name": "perpage",
            "in": "query",
            "description": "Value for perpage",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": 4
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Value for page",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": 0
            }
          },
          {
            "name": "fechaDesde",
            "in": "query",
            "description": "Value for date",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "20/11/2022-13:37:46"
            }
          },
          {
            "name": "fechaHasta",
            "in": "query",
            "description": "Value for date",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "20/12/2022-13:37:46"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTableUplinks"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/uplinks/exportValues": {
      "get": {
        "tags": [
          "uplinks"
        ],
        "summary": "get info for uplinks",
        "description": "get info for uplinks",
        "operationId": "GetExportUplinks",
        "parameters": [
          {
            "name": "Entity",
            "in": "query",
            "description": "Value for entity",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "3ra. Bomba"
            }
          },
          {
            "name": "State",
            "in": "query",
            "description": "Value for state",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "activate"
            }
          },
          {
            "name": "fechaDesde",
            "in": "query",
            "description": "Value for date",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "20/11/2022-13:37:46"
            }
          },
          {
            "name": "fechaHasta",
            "in": "query",
            "description": "Value for date",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "20/12/2022-13:37:46"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTableUplinks"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddPrimaryGroup": {
        "required": [
          "action",
          "NamePrimaryGroup",
          "IdGroup",
          "SecondaryGroups"
        ],
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "example": "add"
          },
          "NamePrimaryGroup": {
            "type": "string",
            "example": "Piscinas-SOP"
          },
          "IdGroup": {
            "type": "string",
            "example": "PSP-1"
          },
          "SecondaryGroups": {
            "type": "object",
            "example": [
              {
                "NameSecondaryGroup": "Pozas-MOP-II",
                "IdSecondaryGroup": "PMP-2",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.120/140",
                    "IdEntity": "BA-120",
                    "Devices": [
                      {
                        "NameDevice": "endev-02",
                        "IdDevice": "EN-2",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Recursos-Salar",
                "IdSecondaryGroup": "RS-3",
                "EntityAsociated": [
                  {
                    "EntityName": "3era.Bomba",
                    "IdEntity": "BA-3",
                    "Devices": [
                      {
                        "NameDevice": "endev-05",
                        "IdDevice": "EN-5",
                        "State": "activate"
                      },
                      {
                        "NameDevice": "endev-06",
                        "IdDevice": "EN-6",
                        "State": "desactivate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Planta-SOP-H",
                "IdSecondaryGroup": "PSH-1",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.TS-6",
                    "IdEntity": "BA-6",
                    "Devices": [
                      {
                        "NameDevice": "endev-07",
                        "IdDevice": "EN-7",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              }
            ]
          }
        }
      },
      "UpdatePrimaryGroup": {
        "required": [
          "NamePrimaryGroup",
          "NewNamePrimaryGroup",
          "IdGroup",
          "NewIdGroup",
          "SecondaryGroups",
          "NewSecondaryGroups"
        ],
        "type": "object",
        "properties": {
          "NamePrimaryGroup": {
            "type": "string",
            "example": "Piscinas-SOP"
          },
          "NewNamePrimaryGroup": {
            "type": "string",
            "example": "Piscinas-SOP-2"
          },
          "IdGroup": {
            "type": "string",
            "example": "PSP-1"
          },
          "NewIdGroup": {
            "type": "string",
            "example": "PSP-2"
          },
          "SecondaryGroups": {
            "type": "object",
            "example": [
              {
                "NameSecondaryGroup": "Pozas-MOP-II",
                "IdSecondaryGroup": "PMP-2",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.120/140",
                    "IdEntity": "BA-120",
                    "Devices": [
                      {
                        "NameDevice": "endev-02",
                        "IdDevice": "EN-2",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Recursos-Salar",
                "IdSecondaryGroup": "RS-3",
                "EntityAsociated": [
                  {
                    "EntityName": "3era.Bomba",
                    "IdEntity": "BA-3",
                    "Devices": [
                      {
                        "NameDevice": "endev-05",
                        "IdDevice": "EN-5",
                        "State": "activate"
                      },
                      {
                        "NameDevice": "endev-06",
                        "IdDevice": "EN-6",
                        "State": "desactivate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Planta-SOP-H",
                "IdSecondaryGroup": "PSH-1",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.TS-6",
                    "IdEntity": "BA-6",
                    "Devices": [
                      {
                        "NameDevice": "endev-07",
                        "IdDevice": "EN-7",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              }
            ]
          },
          "NewSecondaryGroups": {
            "type": "object",
            "example": [
              {
                "NameSecondaryGroup": "Pozas-MOP-II",
                "IdSecondaryGroup": "PMP-2",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.120/140",
                    "IdEntity": "BA-120",
                    "Devices": [
                      {
                        "NameDevice": "endev-02",
                        "IdDevice": "EN-2",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Recursos-Salar",
                "IdSecondaryGroup": "RS-3",
                "EntityAsociated": [
                  {
                    "EntityName": "3era.Bomba",
                    "IdEntity": "BA-3",
                    "Devices": [
                      {
                        "NameDevice": "endev-05",
                        "IdDevice": "EN-5",
                        "State": "activate"
                      },
                      {
                        "NameDevice": "endev-06",
                        "IdDevice": "EN-6",
                        "State": "desactivate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Planta-SOP-H",
                "IdSecondaryGroup": "PSH-1",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.TS-6",
                    "IdEntity": "BA-6",
                    "Devices": [
                      {
                        "NameDevice": "endev-07",
                        "IdDevice": "EN-7",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              }
            ]
          }
        }
      },
...
      "ResponseAddPrimaryGroup": {
        "required": [
          "action",
          "NamePrimaryGroup",
          "IdGroup",
          "SecondaryGroups"
        ],
        "type": "object",
        "properties": {
          "NamePrimaryGroup": {
            "type": "string",
            "example": "Piscinas-SOP"
          },
          "IdGroup": {
            "type": "string",
            "example": "PSP-1"
          },
          "StatusGroup": {
            "type": "string",
            "example": "added"
          },
          "SecondaryGroups": {
            "type": "object",
            "example": [
              {
                "NameSecondaryGroup": "Pozas-MOP-II",
                "IdSecondaryGroup": "PMP-2",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.120/140",
                    "IdEntity": "BA-120",
                    "Devices": [
                      {
                        "NameDevice": "endev-02",
                        "IdDevice": "EN-2",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Recursos-Salar",
                "IdSecondaryGroup": "RS-3",
                "EntityAsociated": [
                  {
                    "EntityName": "3era.Bomba",
                    "IdEntity": "BA-3",
                    "Devices": [
                      {
                        "NameDevice": "endev-05",
                        "IdDevice": "EN-5",
                        "State": "activate"
                      },
                      {
                        "NameDevice": "endev-06",
                        "IdDevice": "EN-6",
                        "State": "desactivate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Planta-SOP-H",
                "IdSecondaryGroup": "PSH-1",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.TS-6",
                    "IdEntity": "BA-6",
                    "Devices": [
                      {
                        "NameDevice": "endev-07",
                        "IdDevice": "EN-7",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              }
            ]
          }
        }
      },
      "ResponseUpdatePrimaryGroup": {
        "required": [
          "NamePrimaryGroup",
          "IdGroup",
          "SecondaryGroups"
        ],
        "type": "object",
        "properties": {
          "NamePrimaryGroup": {
            "type": "string",
            "example": "Piscinas-SOP"
          },
          "IdGroup": {
            "type": "string",
            "example": "PSP-1"
          },
          "StatusGroup": {
            "type": "string",
            "example": "updated"
          },
          "SecondaryGroups": {
            "type": "object",
            "example": [
              {
                "NameSecondaryGroup": "Pozas-MOP-II",
                "IdSecondaryGroup": "PMP-2",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.120/140",
                    "IdEntity": "BA-120",
                    "Devices": [
                      {
                        "NameDevice": "endev-02",
                        "IdDevice": "EN-2",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Recursos-Salar",
                "IdSecondaryGroup": "RS-3",
                "EntityAsociated": [
                  {
                    "EntityName": "3era.Bomba",
                    "IdEntity": "BA-3",
                    "Devices": [
                      {
                        "NameDevice": "endev-05",
                        "IdDevice": "EN-5",
                        "State": "activate"
                      },
                      {
                        "NameDevice": "endev-06",
                        "IdDevice": "EN-6",
                        "State": "desactivate"
                      }
                    ]
                  }
                ]
              },
              {
                "NameSecondaryGroup": "Planta-SOP-H",
                "IdSecondaryGroup": "PSH-1",
                "EntityAsociated": [
                  {
                    "EntityName": "Bba.TS-6",
                    "IdEntity": "BA-6",
                    "Devices": [
                      {
                        "NameDevice": "endev-07",
                        "IdDevice": "EN-7",
                        "State": "activate"
                      }
                    ]
                  }
                ]
              }
            ]
          }
        }
      },
      "ResponseConsultOrganization": {
        "type": "object",
        "properties": {
          "Response": {
            "type": "object",
            "example": [
              {
                "NamePrimaryGroup": "Campamento-andino",
                "IdGroup": "CA-1",
                "SecondaryGroups": [
                  {
                    "NameSecondaryGroup": "Recursos-Salar",
                    "IdSecondaryGroup": "RS-1",
                    "EntityAsociated": [
                      {
                        "EntityName": "3era.Bomba",
                        "IdEntity": "BA-3",
                        "Devices": [
                          {
                            "NameDevice": "endev-01",
                            "IdDevice": "EN-1",
                            "State": "activate"
                          }
                        ]
                      }
                    ]
                  }
                ]
              },
              {
                "NamePrimaryGroup": "Piscinas-SOP",
                "IdGroup": "PSP-1",
                "SecondaryGroups": [
                  {
                    "NameSecondaryGroup": "Pozas-MOP-II",
                    "IdSecondaryGroup": "PMP-2",
                    "EntityAsociated": [
                      {
                        "EntityName": "Bba.120/140",
                        "IdEntity": "BA-120",
                        "Devices": [
                          {
                            "NameDevice": "endev-02",
                            "IdDevice": "EN-2",
                            "State": "activate"
                          }
                        ]
                      }
                    ]
                  },
                  {
                    "NameSecondaryGroup": "Recursos-Salar",
                    "IdSecondaryGroup": "RS-3",
                    "EntityAsociated": [
                      {
                        "EntityName": "3era.Bomba",
                        "IdEntity": "BA-3",
                        "Devices": [
                          {
                            "NameDevice": "endev-05",
                            "IdDevice": "EN-5",
                            "State": "activate"
                          },
                          {
                            "NameDevice": "endev-06",
                            "IdDevice": "EN-6",
                            "State": "desactivate"
                          }
                        ]
                      }
                    ]
                  },
                  {
                    "NameSecondaryGroup": "Planta-SOP-H",
                    "IdSecondaryGroup": "PSH-1",
                    "EntityAsociated": [
                      {
                        "EntityName": "Bba.TS-6",
                        "IdEntity": "BA-6",
                        "Devices": [
                          {
                            "NameDevice": "endev-07",
                            "IdDevice": "EN-7",
                            "State": "activate"
                          }
                        ]
                      }
                    ]
                  }
                ]
              },
              {
                "NamePrimaryGroup": "Piscinas-SOP-2",
                "IdGroup": "PSP-1",
                "SecondaryGroups": [
                  {
                    "NameSecondaryGroup": "Pozas-MOP-II",
                    "IdSecondaryGroup": "PMP-2",
                    "EntityAsociated": [
                      {
                        "EntityName": "Bba.120/140",
                        "IdEntity": "BA-120",
                        "Devices": [
                          {
                            "NameDevice": "endev-02",
                            "IdDevice": "EN-2",
                            "State": "activate"
                          }
                        ]
                      }
                    ]
                  },
                  {
                    "NameSecondaryGroup": "Recursos-Salar",
                    "IdSecondaryGroup": "RS-3",
                    "EntityAsociated": [
                      {
                        "EntityName": "3era.Bomba",
                        "IdEntity": "BA-3",
                        "Devices": [
                          {
                            "NameDevice": "endev-05",
                            "IdDevice": "EN-5",
                            "State": "activate"
                          },
                          {
                            "NameDevice": "endev-06",
                            "IdDevice": "EN-6",
                            "State": "desactivate"
                          }
                        ]
                      }
                    ]
                  },
                  {
                    "NameSecondaryGroup": "Planta-SOP-H",
                    "IdSecondaryGroup": "PSH-1",
                    "EntityAsociated": [
                      {
                        "EntityName": "Bba.TS-6",
                        "IdEntity": "BA-6",
                        "Devices": [
                          {
                            "NameDevice": "endev-07",
                            "IdDevice": "EN-7",
                            "State": "activate"
                          }
                        ]
                      }
                    ]
                  }
                ]
              }
            ]
          }
        }
      },
      "ResponseDeleteGroup": {
        "required": [
          "NamePrimaryGroup",
          "StatusGroup"
        ],
        "type": "object",
        "properties": {
          "NamePrimaryGroup": {
            "type": "string",
            "example": "Piscina-MOP"
          },
          "StatusGroup": {
            "type": "string",
            "example": "removed"
          }
        }
      },
...
    }
  }
}